+++ CKYazdirDb/Model/Kutuk.cs
@@ -13,7 +13,7 @@ namespace ODM.CKYazdirDb
     {
         [Key, DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public int Id { get; set; }
-        public int OpaqId { get; set; }
+        public long OpaqId { get; set; }
         [StringLength(50)]
         public string IlAdi { get; set; }
         [StringLength(50)]
